requirements.txt
appium-python-client
pytest
loguru